requirements.txt
Django==1.8.4
djangorestframework==3.6.3
MySQL-python==1.2.5
Pillow==3.3.1
pyOpenSSL==0.13.1
requests==2.6.0